requirements_legacy.txt
# Legacy requirements for Python 3.8+ (older versions)
quotexpy==1.0.38
telethon>=1.24.0,<1.35.0
requests>=2.25.0,<3.0.0
python-dotenv>=0.19.0,<2.0.0
colorama>=0.4.4,<1.0.0
termcolor>=1.1.0,<3.0.0
pytz>=2021.1
pydantic>=1.8.0,<2.0.0
aiohttp>=3.7.0,<4.0.0
websockets>=10.0,<12.0
aiofiles>=0.8.0,<1.0.0
cryptography>=3.4.8
schedule>=1.1.0,<2.0.0
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<2.0.0
sqlalchemy>=1.4.0,<2.0.0